product store tests
I plan to drop the support of DETS because it would implies many things
to make it viable, such as sharding and such.
Because the 2gb limit size is valid for this exercise, we can easily
make a projection of what to do if we go beyond 2gb. And it makes things
harder.