add a cart registry
i've decided to go for a registry to handle carts. the most obvious
reason is that I just want to try it. I've already used it as a caching
system. also wanted to dip my toes with the match spectification syntax
provided by Erlang as you can see on `list_carts/0` function.