extracted price configuration
I was not very happy with the hardcoded rules of the price
configuration.
So i've decided to extract into a single file that can easily be updated
It worth noting that because it's now a config, we've to reboot the app
in order to taking into account a fresh version of it.